Added redirected paths for login, logout, and register. Spearated logout logic in own component and render according to whether you are logging in or not
