+++ README.md
@@ -1,2 +1,11 @@
 # Uniroute: Map Routing System for Students
 The project has been created by: Chandler Dugan, Christian Hart, and Eric Rivas
+## Initial Setup
+### Before you get started, you need to install some dependencies in your environment. 
+1. First open up two terminals in your coding environment
+2. In your first terminal, navigate to the api folder and create your python virtual environment
+3. In your second terminal make sure you have node and npm installed
+4. Activate your python environmet in the first Terminal and type in this install command: 'pip install flask flask-mysqldb python-dotenv googlemaps'
+5. In the other terminal type in this install command: 'npm install react react-icons react-router-dom axios google-maps-react' 
+6. After that in your python environment start you backend with 'flask run' and in the other terminal start up react with 'npm start' 
+## When using the app
