+++ api/api.py
@@ -87,6 +87,7 @@ def create_token():
     access_token = create_access_token(identity=username)
     response = {"access_token":access_token}
     return response
+
 # this refreshes the jwt authentication so it does not randomly log you out
 @app.after_request
 def refresh_expiring_jwts(response):
@@ -105,7 +106,7 @@ def refresh_expiring_jwts(response):
         # case where there is not a valid JWT. Just return the original response
         return response
 
-# Api route for loggin out users
+# Api route for logging out users
 @app.route('/logout', methods=['POST'])
 def logout():
     response = jsonify({"msg": "logout successful"})
@@ -184,6 +185,12 @@ def get_google_route():
     # print(get_route().text)
     return get_route()
 
-@app.route('/schedule')
+@app.route('/get_schedule')
 def get_schedule():
-    return "PLACEHOLDER FOR MARTA TRAIN DATA"
+    return "PLACE HOLDER FOR GETTING USER SCHEDULE"
+
+# Will need to talk to see if we should store by day or be able to store the whole week if you choose to do so
+# Leaning towards imputing individual days / events similar to outlook & apple calendarss
+@app.route('/save_schedule')
+def save_schedule():
+    return "PLACE HOLDER FOR STORiNG SCHEDUKE"
